+++ gwcelery/tasks/superevents.py
@@ -19,13 +19,11 @@ log = get_task_logger(__name__)
 @lvalert.handler('cbc_gstlal',
                  'cbc_pycbc',
                  'cbc_mbtaonline',
-                 'cbc_gstlal_mdc',
                  'burst_lib',
                  'burst_cwb',
                  'test_gstlal',
                  'test_pycbc',
                  'test_mbtaonline',
-                 'test_gstlal_mdc',
                  queue='superevent',
                  shared=False)
 def handle(payload):